+++ api/exp/swagger/swagger.pb.go
@@ -332,23 +332,6 @@ Livelesson = `{
           "LiveLessonsService"
         ]
       }
-    },
-    "/exp/livelessonall": {
-      "get": {
-        "summary": "Retrieve all livelessons",
-        "operationId": "ListLiveLessons",
-        "responses": {
-          "200": {
-            "description": "",
-            "schema": {
-              "$ref": "#/definitions/expLiveLessonMap"
-            }
-          }
-        },
-        "tags": [
-          "LiveLessonsService"
-        ]
-      }
     }
   },
   "definitions": {
@@ -388,17 +371,6 @@ Livelesson = `{
       "properties": {
         "id": {
           "type": "string"
-        }
-      }
-    },
-    "expLessontoUUIDMap": {
-      "type": "object",
-      "properties": {
-        "Uuids": {
-          "type": "object",
-          "additionalProperties": {
-            "$ref": "#/definitions/expUUIDtoLiveLessonMap"
-          }
         }
       }
     },
@@ -455,9 +427,6 @@ Livelesson = `{
           "type": "string",
           "format": "date-time"
         },
-        "sessionId": {
-          "type": "string"
-        },
         "LessonDiagram": {
           "type": "string"
         },
@@ -471,25 +440,15 @@ Livelesson = `{
       },
       "description": "A provisioned lab without the scheduler details. The server will translate from an underlying type\n(i.e. KubeLab) into this, so only the abstract, relevant details are presented."
     },
-    "expLiveLessonMap": {
-      "type": "object",
-      "properties": {
-        "Sessions": {
-          "type": "object",
-          "additionalProperties": {
-            "$ref": "#/definitions/expLessontoUUIDMap"
-          }
-        }
-      }
-    },
-    "expUUIDtoLiveLessonMap": {
+    "expSyringeState": {
       "type": "object",
       "properties": {
         "Livelessons": {
           "type": "object",
           "additionalProperties": {
             "$ref": "#/definitions/expLiveLesson"
-          }
+          },
+          "title": "Map that contains a mapping of UUIDs to LiveLesson messages"
         }
       }
     }